% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_agreement}
\alias{agr_get_agreement}
\title{Get a particular agreement general details}
\usage{
agr_get_agreement(id_agreement)
}
\arguments{
\item{id_agreement}{A hashed id of NFZ agreement.}
}
\value{
Returns data.table containing data about particular agreements.
}
\description{
Query an agreement schema from agreements API for contract's number of units, prices and average price split by products.
}
