% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{agr_get_agreements}
\alias{agr_get_agreements}
\title{Get all agreements meeting criteria}
\usage{
agr_get_agreements(year, admin_branch, ...)
}
\arguments{
\item{year}{A year of agreements.}

\item{admin_branch}{The number of one of 16 voivodships branches of NFZ; character with leading zero between 02-32, increased by 2.}

\item{...}{One or multiple argument to filter providers.}
}
\value{
Returns data.table.
}
\description{
Query an agreement schema from agreements API to retrieve data about NFZ agreements meeting specified arguments.
}
\examples{
\dontrun{
 get_agreements(year=2021, admin_branch='07', town='Siedlce')
}
}
