% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{agr_get_month_plan}
\alias{agr_get_month_plan}
\title{Get an agreement particular month's plan}
\usage{
agr_get_month_plan(id_plan)
}
\arguments{
\item{id_plan}{A hashed id of a month plan of a NFZ agreement.}
}
\value{
Returns data.table.
}
\description{
Query an agreement schema from agreements API for data about a particular agreement month plan split by products. Data contains number of units, price and average price by products.
}
