% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_plan}
\alias{agr_get_plan}
\title{Get a particular agreement general data by months}
\usage{
agr_get_plan(id_agreement)
}
\arguments{
\item{id_agreement}{A hashed id of a NFZ agreement.}
}
\value{
Returns data.table.
}
\description{
Query an agreement schema from agreements API for data about about a particular agreement split by months. Data contains number of units, price and average price without.
}
