% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{agr_get_prov_by_year}
\alias{agr_get_prov_by_year}
\title{Get a list of providers from a year}
\usage{
agr_get_prov_by_year(year, ...)
}
\arguments{
\item{year}{A year of a contract.}

\item{...}{One or multiple argument to filter providers.}
}
\value{
Returns data.table.
}
\description{
Query a providers schema from agreements API for data about providers having contract with NFZ in a particular year and meeting any other specified with arguments criteria. A table contains contract's amount.
}
