% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_provider}
\alias{agr_get_provider}
\title{Get particular provider's contracted services}
\usage{
agr_get_provider(year, provider_code, admin_branch)
}
\arguments{
\item{year}{A year of a contract.}

\item{provider_code}{NFZ assigned provider code (KOD_SWIADCZENIODAWCY).}

\item{admin_branch}{NFZ voivodship branch code (OW).}
}
\value{
Returns data.table.
}
\description{
Query a providers schema from agreements API for data about a provider's NFZ contracted services in a particular year.
}
