% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{agr_get_providers}
\alias{agr_get_providers}
\title{Get a list of providers}
\usage{
agr_get_providers(...)
}
\arguments{
\item{...}{One or multiple argument to filter providers.}
}
\value{
Returns data.table.
}
\description{
Query a providers schema from agreements API for data about providers with contracts with NFZ meeting with arguments criteria.
}
