% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{agr_get_serivces}
\alias{agr_get_serivces}
\title{Get a list of service types (Rodzaje Swiadczen)}
\usage{
agr_get_serivces(year, service_type, service_name)
}
\arguments{
\item{year}{A year of a contract.}

\item{service_type}{A full or part of NFZ a service type code (Kod rodzaju swiadczen)}

\item{service_name}{A full or part of NFZ a service type name (Nazwa rodzaju swiadczen)}
}
\value{
Returns data.table.
}
\description{
Query a service-types schema from agreements API to get a list of service types (Rodzaje Swiadczen) available in a given year.
}
