% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_serivces_year}
\alias{agr_get_serivces_year}
\title{Get service types available for a given year}
\usage{
agr_get_serivces_year(year)
}
\arguments{
\item{year}{a year for valid services}
}
\value{
Returns a data.table
}
\description{
Function to retrieve all service types valid for a given year
}
