% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_years}
\alias{agr_get_years}
\title{Get years of available data}
\usage{
agr_get_years()
}
\value{
Returns integer vector
}
\description{
Function to retrieve a all years for which data is available
}
