% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_req_args}
\alias{check_req_args}
\title{Check for function's arguments}
\usage{
check_req_args(gen = 0)
}
\arguments{
\item{gen}{Indicates on how many environments up is the function call which arguments are to be checked, relative to where check_req_args is called at. Equivalent to sys.parent's 'n' argument.}
}
\value{
Returns a vector of arguments provided in a function call if all arguments are correct, otherwise function stops
}
\description{
To be run in local environment (inside another function). Checks whether all required arguments of function run in global env were provided during a function call. If a vector of possible arguments was given also checks whether all provided in a function call arguments are supported.
}
\examples{
\dontrun{
 foo <- function(a, b, ...){
   given_args = check_req_args(foo, c("a", "b", "c"))
   return(given_args)}

 foo("a", "b", "c", "d")
}

}
