% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{convert_pl_signs}
\alias{convert_pl_signs}
\title{Convert Polish non ASCII characters to NFZapi accpted utf-8 representation of Unicode}
\usage{
convert_pl_signs(char)
}
\arguments{
\item{char}{A character where the signs are to be converted}
}
\value{
Returns a character vector
}
\description{
Function accpets a string and converts Polish diacretic non ASCII characters
}
\examples{
check_req_args(a, b) <- example to run
\dontrun{
 convert_pl_signs("Gąśłękówiżań")
}

}
