% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_func.R
\name{get_request}
\alias{get_request}
\title{Get data from selected api type and schema provided available arguments}
\usage{
get_request(
  available_args,
  api_type = "app-umw-api",
  schema = "agreements",
  url_args = NULL
)
}
\arguments{
\item{available_args}{Dictionary of function arguments and their corresponding equivalents in NFZ API.}

\item{api_type}{A type of NFZ api to be queried.}

\item{schema}{A name of schema of appropriate api.}

\item{url_args}{A vector of obligatory arguments to be passed without keys as a part of the url \url{https://api.nfz.gov.pl/api_type/schema/url_args/url_args}.}
}
\value{
Returns a data.table for requested query.
}
\description{
General function to get data from selected NFZ API type (agreements, ...) and schema (ex. agreements or providers), provided list of available arguments. Function to be run inside outer function for selected API User given arguments are retrieved from appropriate environment.
}
