% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_agreements}
\alias{agr_get_agreements}
\title{Get all agreements meeting criteria}
\usage{
agr_get_agreements(
  year,
  admin_branch,
  service_type = NULL,
  product_code = NULL,
  provider_code = NULL,
  provider_name = NULL,
  town = NULL,
  nip = NULL,
  regon = NULL
)
}
\arguments{
\item{year}{A year of agreements.}

\item{admin_branch}{The number of one of 16 voivodships branches of NFZ; character with leading zero between 02-32, increased by 2.}

\item{service_type}{A part of or a full NFZ service type code (kod rodzaju swiadczen)}

\item{product_code}{A part of or a full NFZ product type code (kod produktu kontraktowego)}

\item{provider_code}{A part of or a full provider's code assigned by NFZ for contracting purposes (kod swiadczeniodawcy)}

\item{provider_name}{A part of or a full provider's name (nazwa swiadczeniodawcy)}

\item{town}{A name of the provider's location town or city}

\item{nip}{A part of or a full provider's NIP code. NIP is a ten-digit code used to identify taxpayers in Poland.}

\item{regon}{A part of or full REGON identification number (National Official Business Register).}
}
\value{
Returns data.table.
}
\description{
Query an agreement schema from agreements API to retrieve data about NFZ agreements meeting specified criteria
}
\examples{
\dontrun{
 get_agreements(year=2021, admin_branch='07', town='Siedlce')
}
}
