% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_agreements}
\alias{agr_get_agreements}
\title{Get all agreements meeting criteria}
\usage{
agr_get_agreements(
  year,
  admin_branch,
  service_type = NULL,
  product_code = NULL,
  provider_code = NULL,
  provider_name = NULL,
  town = NULL,
  regon = NULL
)
}
\arguments{
\item{year}{A year of agreements.}

\item{admin_branch}{The number of one of 16 voivodships branches of NFZ; character type, a number in a range of 01-16 with leading zero.}

\item{service_type}{Full NFZ service type code (kod rodzaju swiadczen)}

\item{product_code}{Full NFZ product type code (kod produktu kontraktowego)}

\item{provider_code}{Part of or full provider's code assigned by NFZ for contracting purposes (kod swiadczeniodawcy)}

\item{provider_name}{Part of or full provider's name (nazwa swiadczeniodawcy)}

\item{town}{A name of the provider's location town or city}

\item{regon}{Part of or full REGON identification number (National Official Business Register).}
}
\value{
Returns data.table.
}
\description{
Query an agreement schema from agreements API to retrieve data about NFZ agreements meeting specified criteria
}
\details{
\code{gen} describes a relative position or a stack number of check_req_args parent environment relative to a global environment (global has number 0). Can't be 0 or negative.
}
\examples{
\dontrun{
 get_agreements(year=2021, admin_branch='07', town='Siedlce')
}
}
