% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_products}
\alias{agr_get_products}
\title{Get a list of product types (Produkty Kontraktowe)}
\usage{
agr_get_products(year, prod_code = NULL, product_name = NULL)
}
\arguments{
\item{year}{A year of a contract.}

\item{prod_code}{Full or part of NFZ a product type code (Kod produktu kontraktowego)}

\item{product_name}{Full or part of NFZ a product type name (Nazwa produktu kontraktowego)}
}
\value{
Returns data.table.
}
\description{
Query a product-types schema from agreements API to get a list of product types (Produkty Kontraktowe) available in a given year.
}
