% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreements_api_func.R
\name{agr_get_prov_by_year}
\alias{agr_get_prov_by_year}
\title{Get a list of providers available in given year}
\usage{
agr_get_prov_by_year(
  year,
  admin_branch = NULL,
  provider_code = NULL,
  provider_name = NULL,
  nip = NULL,
  regon = NULL,
  post_code = NULL,
  street = NULL,
  town = NULL,
  teryt = NULL
)
}
\arguments{
\item{year}{A year of agreements.}

\item{admin_branch}{The number of one of 16 voivodships branches of NFZ; character type, a number in a range of 01-16 with leading zero.}

\item{provider_code}{Part of or full provider's code assigned by NFZ for contracting purposes (kod swiadczeniodawcy).}

\item{provider_name}{Part of or full provider's name (nazwa swiadczeniodawcy).}

\item{nip}{Part of or full provider's NIP code. NIP is a ten-digit code used to identify taxpayers in Poland.}

\item{regon}{Part of or full REGON identification number (National Official Business Register).}

\item{post_code}{Postal code of a provider's main location (headquarter). Given without a dash, just 5 digits.}

\item{street}{Street name of provider's main location (headquarter).}

\item{town}{A name of the provider's location town or city.}

\item{teryt}{Location code (TREC) of the National Official Register of the Territorial Division of the Country (TERYT).}
}
\value{
Returns data.table.
}
\description{
Query providers schema from agreements API for data about providers having contract with NFZ in a particular year and meeting any other specified criteria. A table contains contract's amount.
}
