% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_func.R
\name{check_req_args}
\alias{check_req_args}
\title{Check for function's arguments}
\usage{
check_req_args(gen)
}
\arguments{
\item{gen}{The stack number of a frame (environment) which arguments are to be verified}
}
\value{
Returns a vector of arguments provided in a function call if all arguments are correct, otherwise function stops
}
\description{
To be run in local environment (inside another function). Checks whether all required arguments of function run in global env were provided during a function call. If a vector of possible arguments was given also checks whether all provided in a function call arguments are supported.
}
\details{
\code{gen} describes a relative position or a stack number of check_req_args parent environment relative to a global environment (global has number 0). Can't be 0 or negative.
}
\examples{
\dontrun{
 foo <- function(a, b, ...){
   given_args = check_req_args(foo, c("a", "b", "c"))
   return(given_args)}

 foo("a", "b", "c", "d")
}

}
