% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_func.R
\name{convert_pl_signs}
\alias{convert_pl_signs}
\title{Convert Polish non ASCII characters to NFZ API accpted utf-8 representation of Unicode}
\usage{
convert_pl_signs(char)
}
\arguments{
\item{char}{A character where the signs are to be converted}
}
\value{
Returns a character vector
}
\description{
Function accpets a string and converts Polish diacretic non ASCII characters
}
\examples{
\dontrun{
 convert_pl_signs("Gąśłękówiżań")
}

}
